refactor(voice/capture): extract Linux-pair RestartMixin methods [T1.4 step 8b]
Append the Linux-pair restart methods + their canonical-name
sibling-discovery helpers to the existing ``RestartMixin`` at
``voice/capture/_restart_mixin.py`` per CLAUDE.md anti-pattern
#16:

* ``request_alsa_hw_direct_restart`` (~200 LOC) — Linux twin of
  ``request_exclusive_restart``: bypass PipeWire / PulseAudio
  filter chains by reopening the kernel ALSA sibling. Returns a
  structured :class:`AlsaHwDirectRestartResult` distinguishing
  ``ALSA_HW_ENGAGED`` from ``DOWNGRADED_TO_SESSION_MANAGER``.
* ``request_session_manager_restart`` (~190 LOC) — Linux twin of
  ``request_shared_restart``: revert to PipeWire/PulseAudio.
  Serves both the ``LinuxPipeWireDirectBypass`` revert path AND
  the ``LinuxSessionManagerEscapeBypass`` apply path (explicit
  ``target_device``).
* ``_find_sibling_with_host_api`` (~10 LOC) — wrapper.
* ``_find_sibling_with_host_api_in`` (~35 LOC) — canonical-name
  sibling lookup honouring DI-provided ``enumerate_fn``.

Total: ~440 LOC migrated. ``_capture_task.py`` reduces from 1611
to 1170 LOC — within the CLAUDE.md anti-pattern #16 ceiling
(~500 LOC mixed-responsibility) for the first time since the
god-file split began.

The mixin now owns the full Windows-pair + Linux-pair restart
surface (5 public ``request_*_restart`` methods + 2 helpers + 1
private reconnect helper). New Linux-pair imports added: ``sys``
plus ``_LINUX_ALSA_HOST_API``, ``_LINUX_SESSION_MANAGER_HOST_APIS``,
the ``AlsaHwDirect*`` and ``SessionManager*`` result/verdict types,
and their two metric emitters.

Gates: ruff + ruff format + mypy strict (427 files) + bandit
clean. Full pytest 11894 passed (one more than step 8a — the
previously-flaky ``test_output_guard::test_none_near_zero``
passed in this run).

Mission: docs-internal/missions/MISSION-voice-final-skype-grade-2026.md §Phase 1.T1.4 step 8b
