+++ src/sovyx/voice/capture/_restart_mixin.py
@@ -7,7 +7,7 @@ behaviour across two files keeps each under the CLAUDE.md anti-
 pattern #16 ceiling (~500 LOC mixed-responsibility) — types own
 shape, mixin owns transactional restart semantics.
 
-Step 8a (this commit) lands the **Windows pair**:
+Step 8a landed the **Windows pair**:
 
 * :meth:`_reopen_stream_after_device_error` — generic reconnect
   helper used by the consumer loop on ``sd.PortAudioError`` AND by
@@ -17,9 +17,19 @@ Step 8a (this commit) lands the **Windows pair**:
   exclusive engagement from a downgraded shared-mode reopen.
 * :meth:`request_shared_restart` — symmetric revert to shared mode.
 
-Step 8b will land the **Linux pair**
-(:meth:`request_alsa_hw_direct_restart`,
-:meth:`request_session_manager_restart`) on the same mixin.
+Step 8b (this commit) lands the **Linux pair** + supporting
+sibling-discovery helpers:
+
+* :meth:`request_alsa_hw_direct_restart` — bypass the PipeWire /
+  PulseAudio session manager by reopening the kernel ALSA sibling
+  directly. Twin of ``request_exclusive_restart`` for Linux.
+* :meth:`request_session_manager_restart` — symmetric revert to
+  the PipeWire/PulseAudio session manager. Also serves the
+  ``LinuxSessionManagerEscapeBypass`` apply path with an explicit
+  ``target_device``.
+* :meth:`_find_sibling_with_host_api` /
+  :meth:`_find_sibling_with_host_api_in` — canonical-name sibling
+  lookup helpers shared by the Linux pair.
 
 Mixin contract — the host class (``AudioCaptureTask``)
 initialises the stream-state attributes in ``__init__``. Method
@@ -35,6 +45,7 @@ from __future__ import annotations
 
 import asyncio
 import contextlib
+import sys
 from typing import TYPE_CHECKING, Any
 
 from sovyx.engine.config import VoiceTuningConfig as _VoiceTuning
@@ -43,11 +54,19 @@ from sovyx.voice._agc2 import build_agc2_if_enabled
 from sovyx.voice._frame_normalizer import FrameNormalizer
 from sovyx.voice.capture._helpers import _resolve_input_entry
 from sovyx.voice.capture._restart import (
+    _LINUX_ALSA_HOST_API,
+    _LINUX_SESSION_MANAGER_HOST_APIS,
+    AlsaHwDirectRestartResult,
+    AlsaHwDirectRestartVerdict,
     ExclusiveRestartResult,
     ExclusiveRestartVerdict,
+    SessionManagerRestartResult,
+    SessionManagerRestartVerdict,
     SharedRestartResult,
     SharedRestartVerdict,
+    _emit_alsa_hw_direct_restart_metric,
     _emit_exclusive_restart_metric,
+    _emit_session_manager_restart_metric,
     _emit_shared_restart_metric,
 )
 
@@ -78,8 +97,11 @@ class RestartMixin:
     so the bypass coordinator can distinguish "engaged" from
     "downgraded" without parsing logs.
 
-    The Linux pair (step 8b: ALSA-hw-direct ↔ session-manager) lands
-    in the same mixin in a follow-up commit.
+    Linux pair (step 8b): ALSA-hw-direct ↔ session-manager toggle
+    for the PipeWire/PulseAudio bypass strategies. Same engaged-
+    vs-downgraded distinction in
+    :class:`AlsaHwDirectRestartResult` /
+    :class:`SessionManagerRestartResult`.
     """
 
     # Host-class state declarations for mypy strict. The host class
@@ -484,3 +506,444 @@ class RestartMixin:
         )
         _emit_shared_restart_metric(result)
         return result
+
+    async def request_alsa_hw_direct_restart(self) -> AlsaHwDirectRestartResult:
+        """Reopen the capture stream against the ALSA-direct sibling device.
+
+        Linux-specific twin of :meth:`request_exclusive_restart`. The
+        ``LinuxPipeWireDirectBypass`` strategy invokes this when it
+        wants to bypass a misbehaving PipeWire/PulseAudio filter chain
+        (e.g. ``module-echo-cancel``, ``rnnoise`` filter, user-added
+        EQ) and talk to the kernel ALSA device directly.
+
+        Resolution: re-enumerate input devices, locate the sibling whose
+        :attr:`DeviceEntry.canonical_name` matches the current endpoint
+        AND whose :attr:`DeviceEntry.host_api_name` equals ``"ALSA"``.
+        When found, that entry is handed to the unified opener as the
+        starting point — the opener's sibling-chain fallback then
+        automatically covers the "ALSA open refused, fall back to
+        PulseAudio" path.
+
+        Idempotent — safe to call while stopped or on a non-Linux host;
+        in either case it is a no-op and the existing stream (if any)
+        is preserved.
+
+        Returns:
+            An :class:`AlsaHwDirectRestartResult`. Callers inspect
+            ``result.engaged`` (``True`` iff the ALSA host API actually
+            won the fallback pyramid) to know whether the PipeWire
+            bypass is in effect.
+        """
+        if not self._running:
+            logger.debug("audio_capture_alsa_hw_direct_restart_skipped_not_running")
+            result = AlsaHwDirectRestartResult(
+                verdict=AlsaHwDirectRestartVerdict.NOT_RUNNING,
+                engaged=False,
+                detail="capture task is not running",
+            )
+            _emit_alsa_hw_direct_restart_metric(result)
+            return result
+        if sys.platform != "linux":
+            logger.debug(
+                "audio_capture_alsa_hw_direct_restart_skipped_not_linux",
+                platform=sys.platform,
+            )
+            result = AlsaHwDirectRestartResult(
+                verdict=AlsaHwDirectRestartVerdict.NOT_LINUX,
+                engaged=False,
+                host_api=self._host_api_name,
+                device=self._input_device,
+                sample_rate=self._sample_rate,
+                detail=f"request_alsa_hw_direct_restart is Linux-only; running on {sys.platform}",
+            )
+            _emit_alsa_hw_direct_restart_metric(result)
+            return result
+
+        alsa_entry = self._find_sibling_with_host_api(_LINUX_ALSA_HOST_API)
+        if alsa_entry is None:
+            logger.warning(
+                "audio_capture_alsa_hw_direct_restart_no_sibling",
+                device=self._input_device,
+                host_api=self._host_api_name,
+            )
+            result = AlsaHwDirectRestartResult(
+                verdict=AlsaHwDirectRestartVerdict.NO_ALSA_SIBLING,
+                engaged=False,
+                host_api=self._host_api_name,
+                device=self._input_device,
+                sample_rate=self._sample_rate,
+                detail=(
+                    "no ALSA-host-API sibling found for current endpoint "
+                    "(PortAudio build without ALSA, or device held exclusive)"
+                ),
+            )
+            _emit_alsa_hw_direct_restart_metric(result)
+            return result
+
+        from sovyx.voice._stream_opener import StreamOpenError, open_input_stream
+
+        tuning = self._tuning if self._tuning is not None else _VoiceTuning()
+        logger.warning(
+            "audio_capture_alsa_hw_direct_restart_begin",
+            device=self._input_device,
+            host_api=self._host_api_name,
+            target_host_api=_LINUX_ALSA_HOST_API,
+            target_device_index=alsa_entry.index,
+        )
+
+        # Tear down the existing (session-manager-backed) stream before
+        # we grab the kernel device — some ALSA drivers reject a second
+        # client even for read-only capture.
+        await asyncio.to_thread(self._close_stream, "alsa_hw_direct_restart")
+        while not self._queue.empty():
+            with contextlib.suppress(asyncio.QueueEmpty):
+                self._queue.get_nowait()
+
+        try:
+            stream, info = await open_input_stream(
+                device=alsa_entry,
+                target_rate=self._sample_rate,
+                blocksize=self._blocksize,
+                callback=self._audio_callback,
+                tuning=tuning,
+                sd_module=self._sd_module,
+                enumerate_fn=self._enumerate_fn,
+                validate_fn=None,
+            )
+        except StreamOpenError as exc:
+            logger.error(
+                "audio_capture_alsa_hw_direct_restart_failed",
+                error=str(exc),
+                device=self._input_device,
+                host_api=self._host_api_name,
+            )
+            # Mirror the exclusive-fallback behaviour: try to recover
+            # the pipeline through shared mode so the user is not left
+            # with a dead stream.
+            try:
+                await self._reopen_stream_after_device_error()
+            except Exception as fallback_exc:  # noqa: BLE001
+                logger.error(
+                    "audio_capture_alsa_hw_direct_fallback_failed",
+                    error=str(fallback_exc),
+                )
+                self._signal_consumer_shutdown()
+                result = AlsaHwDirectRestartResult(
+                    verdict=AlsaHwDirectRestartVerdict.OPEN_FAILED_NO_STREAM,
+                    engaged=False,
+                    host_api=self._host_api_name,
+                    device=self._input_device,
+                    detail=(
+                        f"ALSA-direct open failed ({exc}); session-manager "
+                        f"fallback also failed ({fallback_exc})"
+                    ),
+                )
+                _emit_alsa_hw_direct_restart_metric(result)
+                return result
+            result = AlsaHwDirectRestartResult(
+                verdict=AlsaHwDirectRestartVerdict.DOWNGRADED_TO_SESSION_MANAGER,
+                engaged=False,
+                host_api=self._host_api_name,
+                device=self._input_device,
+                sample_rate=self._sample_rate,
+                detail=f"ALSA-direct open failed ({exc}); recovered via session manager",
+            )
+            _emit_alsa_hw_direct_restart_metric(result)
+            return result
+
+        self._stream = stream
+        self._sample_rate = info.sample_rate
+        self._input_device = info.device_index
+        self._host_api_name = info.host_api
+        self._resolved_device_name = alsa_entry.name
+        # F5/F6: AGC2 default-on per VoiceTuningConfig.agc2_enabled
+        # (commit 2e36893). Operators can revert via
+        # SOVYX_TUNING__VOICE__AGC2_ENABLED=false. The factory
+        # returns None when disabled — FrameNormalizer accepts None
+        # as the no-op default so the call site needs no ``if`` branch.
+        _agc2_tuning = self._tuning if self._tuning is not None else _VoiceTuning()
+        self._normalizer = FrameNormalizer(
+            source_rate=info.sample_rate,
+            source_channels=info.channels,
+            agc2=build_agc2_if_enabled(
+                enabled=_agc2_tuning.agc2_enabled,
+                sample_rate=info.sample_rate,
+            ),
+        )
+        self._allocate_ring_buffer(tuning)
+        self._emit_stream_opened(info, apo_bypass_attempted=True)
+
+        if info.host_api != _LINUX_ALSA_HOST_API:
+            logger.error(
+                "audio_capture_alsa_hw_direct_restart_downgraded_to_session_manager",
+                device=self._input_device,
+                host_api=info.host_api,
+                sample_rate=self._sample_rate,
+                channels=info.channels,
+            )
+            result = AlsaHwDirectRestartResult(
+                verdict=AlsaHwDirectRestartVerdict.DOWNGRADED_TO_SESSION_MANAGER,
+                engaged=False,
+                host_api=self._host_api_name,
+                device=self._input_device,
+                sample_rate=self._sample_rate,
+                detail=(
+                    f"opener fell back to {info.host_api!r} — session manager still in signal path"
+                ),
+            )
+            _emit_alsa_hw_direct_restart_metric(result)
+            return result
+        logger.warning(
+            "audio_capture_alsa_hw_direct_restart_ok",
+            device=self._input_device,
+            host_api=self._host_api_name,
+            sample_rate=self._sample_rate,
+            channels=info.channels,
+        )
+        result = AlsaHwDirectRestartResult(
+            verdict=AlsaHwDirectRestartVerdict.ALSA_HW_ENGAGED,
+            engaged=True,
+            host_api=self._host_api_name,
+            device=self._input_device,
+            sample_rate=self._sample_rate,
+        )
+        _emit_alsa_hw_direct_restart_metric(result)
+        return result
+
+    async def request_session_manager_restart(
+        self,
+        target_device: DeviceEntry | None = None,
+    ) -> SessionManagerRestartResult:
+        """Revert the capture stream to the PipeWire/PulseAudio session manager.
+
+        Linux-specific twin of :meth:`request_shared_restart`. Two
+        legitimate callers:
+
+        * :class:`LinuxPipeWireDirectBypass` (revert path) — no
+          ``target_device`` supplied, the method searches for the
+          first sibling whose :attr:`DeviceEntry.host_api_name` lies
+          in :data:`_LINUX_SESSION_MANAGER_HOST_APIS`.
+        * :class:`LinuxSessionManagerEscapeBypass` (apply path, T6
+          of voice-linux-cascade-root-fix) — supplies a concrete
+          ``target_device`` resolved to a session-manager virtual
+          (``pipewire``, ``pulse``) or the OS default ``default`` PCM.
+          The method skips sibling discovery and opens directly.
+
+        When neither path yields a target the method returns
+        :attr:`SessionManagerRestartVerdict.DOWNGRADED_TO_ALSA_HW` with
+        the existing stream preserved.
+
+        Args:
+            target_device: Optional explicit target. When ``None``,
+                the canonical-name-sibling discovery runs. When
+                provided, the method opens against that device
+                verbatim — callers are responsible for pre-filtering.
+
+        Returns:
+            A :class:`SessionManagerRestartResult`. A non-engaged
+            verdict means either the session-manager reopen was not
+            feasible (``DOWNGRADED_TO_ALSA_HW``, ``NO_TARGET``) or the
+            pipeline is now without a live capture
+            (``OPEN_FAILED_NO_STREAM``).
+        """
+        if not self._running:
+            logger.debug("audio_capture_session_manager_restart_skipped_not_running")
+            result = SessionManagerRestartResult(
+                verdict=SessionManagerRestartVerdict.NOT_RUNNING,
+                engaged=False,
+                detail="capture task is not running",
+            )
+            _emit_session_manager_restart_metric(result)
+            return result
+        if sys.platform != "linux":
+            logger.debug(
+                "audio_capture_session_manager_restart_skipped_not_linux",
+                platform=sys.platform,
+            )
+            result = SessionManagerRestartResult(
+                verdict=SessionManagerRestartVerdict.NOT_LINUX,
+                engaged=False,
+                host_api=self._host_api_name,
+                device=self._input_device,
+                sample_rate=self._sample_rate,
+                detail=(
+                    f"request_session_manager_restart is Linux-only; running on {sys.platform}"
+                ),
+            )
+            _emit_session_manager_restart_metric(result)
+            return result
+
+        if target_device is not None:
+            session_entry: DeviceEntry | None = target_device
+        else:
+            session_entry = self._find_sibling_with_host_api_in(
+                _LINUX_SESSION_MANAGER_HOST_APIS,
+            )
+        if session_entry is None:
+            logger.warning(
+                "audio_capture_session_manager_restart_no_sibling",
+                device=self._input_device,
+                host_api=self._host_api_name,
+            )
+            result = SessionManagerRestartResult(
+                verdict=SessionManagerRestartVerdict.DOWNGRADED_TO_ALSA_HW,
+                engaged=False,
+                host_api=self._host_api_name,
+                device=self._input_device,
+                sample_rate=self._sample_rate,
+                detail=(
+                    "no PulseAudio/PipeWire sibling available — device is "
+                    "ALSA-direct only; existing stream preserved"
+                ),
+            )
+            _emit_session_manager_restart_metric(result)
+            return result
+
+        from sovyx.voice._stream_opener import StreamOpenError, open_input_stream
+
+        tuning = self._tuning if self._tuning is not None else _VoiceTuning()
+        logger.warning(
+            "audio_capture_session_manager_restart_begin",
+            device=self._input_device,
+            host_api=self._host_api_name,
+            target_host_api=session_entry.host_api_name,
+            target_device_index=session_entry.index,
+        )
+
+        await asyncio.to_thread(self._close_stream, "session_manager_restart")
+        while not self._queue.empty():
+            with contextlib.suppress(asyncio.QueueEmpty):
+                self._queue.get_nowait()
+
+        try:
+            stream, info = await open_input_stream(
+                device=session_entry,
+                target_rate=self._sample_rate,
+                blocksize=self._blocksize,
+                callback=self._audio_callback,
+                tuning=tuning,
+                sd_module=self._sd_module,
+                enumerate_fn=self._enumerate_fn,
+                validate_fn=None,
+            )
+        except StreamOpenError as exc:
+            logger.error(
+                "audio_capture_session_manager_restart_failed",
+                error=str(exc),
+                device=self._input_device,
+                host_api=self._host_api_name,
+            )
+            self._signal_consumer_shutdown()
+            result = SessionManagerRestartResult(
+                verdict=SessionManagerRestartVerdict.OPEN_FAILED_NO_STREAM,
+                engaged=False,
+                host_api=self._host_api_name,
+                device=self._input_device,
+                detail=f"session-manager reopen failed: {exc}",
+            )
+            _emit_session_manager_restart_metric(result)
+            return result
+
+        self._stream = stream
+        self._sample_rate = info.sample_rate
+        self._input_device = info.device_index
+        self._host_api_name = info.host_api
+        self._resolved_device_name = session_entry.name
+        # F5/F6: AGC2 default-on per VoiceTuningConfig.agc2_enabled
+        # (commit 2e36893). Operators can revert via
+        # SOVYX_TUNING__VOICE__AGC2_ENABLED=false. The factory
+        # returns None when disabled — FrameNormalizer accepts None
+        # as the no-op default so the call site needs no ``if`` branch.
+        _agc2_tuning = self._tuning if self._tuning is not None else _VoiceTuning()
+        self._normalizer = FrameNormalizer(
+            source_rate=info.sample_rate,
+            source_channels=info.channels,
+            agc2=build_agc2_if_enabled(
+                enabled=_agc2_tuning.agc2_enabled,
+                sample_rate=info.sample_rate,
+            ),
+        )
+        self._allocate_ring_buffer(tuning)
+        self._emit_stream_opened(info, apo_bypass_attempted=False)
+
+        if info.host_api not in _LINUX_SESSION_MANAGER_HOST_APIS:
+            logger.warning(
+                "audio_capture_session_manager_restart_downgraded_to_alsa_hw",
+                device=self._input_device,
+                host_api=info.host_api,
+            )
+            result = SessionManagerRestartResult(
+                verdict=SessionManagerRestartVerdict.DOWNGRADED_TO_ALSA_HW,
+                engaged=False,
+                host_api=self._host_api_name,
+                device=self._input_device,
+                sample_rate=self._sample_rate,
+                detail=(
+                    f"opener fell back to {info.host_api!r} — session manager not in signal path"
+                ),
+            )
+            _emit_session_manager_restart_metric(result)
+            return result
+        logger.warning(
+            "audio_capture_session_manager_restart_ok",
+            device=self._input_device,
+            host_api=self._host_api_name,
+            sample_rate=self._sample_rate,
+            channels=info.channels,
+        )
+        result = SessionManagerRestartResult(
+            verdict=SessionManagerRestartVerdict.SESSION_MANAGER_ENGAGED,
+            engaged=True,
+            host_api=self._host_api_name,
+            device=self._input_device,
+            sample_rate=self._sample_rate,
+        )
+        _emit_session_manager_restart_metric(result)
+        return result
+
+    def _find_sibling_with_host_api(self, host_api: str) -> DeviceEntry | None:
+        """Return the enumeration sibling of the current endpoint on ``host_api``.
+
+        Siblings share the same :attr:`DeviceEntry.canonical_name` but
+        are served by different host APIs — on Linux a single USB
+        microphone typically appears once via ``ALSA``, once via
+        ``PulseAudio``, once via ``PipeWire``. Returns ``None`` when the
+        current entry has no sibling on the requested host API.
+        """
+        return self._find_sibling_with_host_api_in(frozenset({host_api}))
+
+    def _find_sibling_with_host_api_in(
+        self,
+        host_apis: frozenset[str],
+    ) -> DeviceEntry | None:
+        """Return the first enumeration sibling whose host API is in ``host_apis``.
+
+        Uses the same :func:`_resolve_input_entry` entry point as the
+        start + restart paths so any DI-provided ``enumerate_fn`` is
+        honoured. Returns ``None`` on enumeration failure rather than
+        raising — the caller translates the absence into a structured
+        verdict.
+        """
+        try:
+            current = _resolve_input_entry(
+                input_device=self._input_device,
+                enumerate_fn=self._enumerate_fn,
+                host_api_name=self._host_api_name,
+            )
+        except RuntimeError:
+            return None
+        canonical = current.canonical_name
+        if self._enumerate_fn is not None:
+            entries = self._enumerate_fn()
+        else:
+            from sovyx.voice.device_enum import enumerate_devices
+
+            entries = enumerate_devices()
+        for entry in entries:
+            if entry.max_input_channels <= 0:
+                continue
+            if entry.canonical_name != canonical:
+                continue
+            if entry.host_api_name in host_apis:
+                return entry
+        return None
